+++ packages/firestore/test/integration/api/README.md
@@ -1,0 +1,4 @@
+All tests in this directory are also run against our (minified)
+firebase-firestore.js build (via 'yarn test:manual' under
+firebase-js-sdk/integration/firestore/) and therefore must not have any
+dependencies on src/ files.
