+++ packages/firestore/test/integration/browser/webchannel.test.ts
@@ -18,7 +18,8 @@ import * as api from '../../../src/protos/firestore_proto_api';
 import { expect } from 'chai';
 import { WebChannelConnection } from '../../../src/platform_browser/webchannel_connection';
 import { DatabaseId, DatabaseInfo } from '../../../src/core/database_info';
-import * as utilHelpers from '../util/helpers';
+import {DEFAULT_PROJECT_ID} from '../util/helpers';
+import {getDefaultDatabaseInfo} from '../util/internal_helpers';
 
 const describeFn = typeof window !== 'undefined' ? describe : xdescribe;
 describeFn('WebChannel', () => {
@@ -42,8 +43,8 @@ describeFn('WebChannel', () => {
   });
 
   it('receives error messages', done => {
-    const projectId = utilHelpers.DEFAULT_PROJECT_ID;
-    const info = utilHelpers.getDefaultDatabaseInfo();
+    const projectId = DEFAULT_PROJECT_ID;
+    const info = getDefaultDatabaseInfo();
     const conn = new WebChannelConnection(info);
     const stream = conn.openStream<api.ListenRequest, api.ListenResponse>(
       'Listen',
